Break messages file into messages package
